implementação módulo fiscal

+++ API/NETCORE/Models/Requests/VendaRequest.cs
@@ -2,6 +2,10 @@
  * Arquivo: API/NETCORE/Models/Requests/VendaRequest.cs
  * Objetivo: define contrato de entrada para operações de registro de vendas e itens do carrinho.
  * Entradas esperadas: recebe dados serializados do frontend nas ações da API.
+ *
+ * Quantity é `decimal` (não `int`) para suportar produtos vendidos por peso/volume (ex.: 0,452
+ * kg) — o frontend já envia esse campo como número JSON puro (nunca foi mascarado como texto
+ * pt-BR), então o model binder do ASP.NET Core aceita o mesmo payload sem mudança de formato.
  */
 namespace HORUSPDV_API.Models.Requests;
 
@@ -19,5 +23,5 @@ public class VendaItemRequest
 {
     public string ProductCode { get; set; } = string.Empty;
     public string ProductName { get; set; } = string.Empty;
-    public int Quantity { get; set; }
+    public decimal Quantity { get; set; }
 }
